std: merge the unix-like io::error modules into one file
The POSIX-like `decode_error_kind` copies under `sys/io/error/` (unix,
wasi, teeos) had drifted apart, each missing different arms. Fold wasi
and teeos into unix: `decode_error_kind` and `is_interrupted`
are shared, while `errno`, `set_errno`, `error_string`, and the
`errno_location` extern stay per-target behind `cfg`, expressed as
exceptions to the unix/wasi/teeos set rather than enumerating the
matching targets, matching the platform `cfg_attr`s the file already had.

teeos now goes through the shared `errno_location` extern instead of
calling `libc::__errno_location()` directly, so it gets the same
`#[ffi_const]` annotation as the other targets. `error_string` is shared
between unix and wasi, with the buffer sized per target: 128 bytes on
unix, where some targets are mindful of stack usage, and 1024 on wasi.

The shared decode table is unix's existing one, so wasi picks up
`ENOTEMPTY` -> `DirectoryNotEmpty` and teeos picks up `EINPROGRESS` ->
`InProgress`, `EMFILE`/`ENFILE` -> `TooManyOpenFiles`, and `EOPNOTSUPP`
-> `Unsupported`, all previously `Uncategorized` on those targets.

No test: with one shared `decode_error_kind` the copies can't drift
again, and no test asserts the errno to `ErrorKind` mappings today.
