+++ library/std/src/sys/io/error/unix.rs
@@ -1,8 +1,15 @@
-use crate::ffi::{CStr, c_char, c_int};
+use crate::ffi::c_int;
+#[cfg(not(target_os = "teeos"))]
+use crate::ffi::{CStr, c_char};
 use crate::io;
 
 unsafe extern "C" {
-    #[cfg(not(any(target_os = "dragonfly", target_os = "vxworks", target_os = "rtems")))]
+    #[cfg(not(any(
+        target_os = "dragonfly",
+        target_os = "vxworks",
+        target_os = "rtems",
+        target_os = "wasi"
+    )))]
     #[cfg_attr(
         any(
             target_os = "linux",
@@ -10,6 +17,7 @@ unsafe extern "C" {
             target_os = "fuchsia",
             target_os = "l4re",
             target_os = "hurd",
+            target_os = "teeos",
         ),
         link_name = "__errno_location"
     )]
@@ -37,7 +45,12 @@ unsafe extern "C" {
 }
 
 /// Returns the platform-specific value of errno
-#[cfg(not(any(target_os = "dragonfly", target_os = "vxworks", target_os = "rtems")))]
+#[cfg(not(any(
+    target_os = "dragonfly",
+    target_os = "vxworks",
+    target_os = "rtems",
+    target_os = "wasi"
+)))]
 #[inline]
 pub fn errno() -> i32 {
     unsafe { (*errno_location()) as i32 }
@@ -45,7 +58,12 @@ pub fn errno() -> i32 {
 
 /// Sets the platform-specific value of errno
 // needed for readdir and syscall!
-#[cfg(all(not(target_os = "dragonfly"), not(target_os = "vxworks"), not(target_os = "rtems")))]
+#[cfg(not(any(
+    target_os = "dragonfly",
+    target_os = "vxworks",
+    target_os = "rtems",
+    target_os = "wasi",
+)))]
 #[allow(dead_code)] // but not all target cfgs actually end up using it
 #[inline]
 pub fn set_errno(e: i32) {
@@ -91,6 +109,25 @@ pub fn set_errno(e: i32) {
 
     unsafe {
         errno = e;
+    }
+}
+
+#[cfg(target_os = "wasi")]
+unsafe extern "C" {
+    #[thread_local]
+    #[link_name = "errno"]
+    static mut libc_errno: libc::c_int;
+}
+
+#[cfg(target_os = "wasi")]
+pub fn errno() -> i32 {
+    unsafe { libc_errno as i32 }
+}
+
+#[cfg(target_os = "wasi")]
+pub fn set_errno(val: i32) {
+    unsafe {
+        libc_errno = val;
     }
 }
 
@@ -153,8 +190,9 @@ pub fn decode_error_kind(errno: i32) -> io::ErrorKind {
 }
 
 /// Gets a detailed string description for the given error number.
+#[cfg(any(target_family = "unix", target_os = "wasi"))]
 pub fn error_string(errno: i32) -> String {
-    const TMPBUF_SZ: usize = 128;
+    const TMPBUF_SZ: usize = if cfg!(target_os = "wasi") { 1024 } else { 128 };
 
     unsafe extern "C" {
         #[cfg_attr(
@@ -185,4 +223,9 @@ pub fn error_string(errno: i32) -> String {
         // it's better to give a low-quality error message than none at all.
         String::from_utf8_lossy(CStr::from_ptr(p).to_bytes()).into()
     }
+}
+
+#[cfg(target_os = "teeos")]
+pub fn error_string(_errno: i32) -> String {
+    "error string unimplemented".to_string()
 }
